+++ packages/vue/types/Link.d.ts
@@ -9,6 +9,7 @@ export interface LinkComponent extends Vue {
     location: HickoryLocation;
     href: string;
     click(e: MouseEvent): void;
+    loading: boolean;
 }
 declare const Link: ComponentOptions<LinkComponent>;
 export default Link;